refactor: restructure CLI and batch download components
- Introduced a new `main.go` file to serve as the entry point for the CLI application.
- Created a `download.go` file to encapsulate batch download logic, improving organization and readability.
- Added comprehensive tests for download functionality and CLI options parsing to ensure reliability.
- Updated the README to include development instructions and clarify usage of the application.
- Removed the old `main_test.go` file as its tests have been migrated to the new structure.
