+++ cmd/yamdl/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"os"
+	"ya-music/internal/cli"
+)
+
+func main() {
+	os.Exit(cli.Run(os.Args[1:], os.Stdout, os.Stderr))
+}
