+++ internal/batch/download_test.go
