+++ internal/batch/format_test.go
